First cut at replacing glpk.cpp with pip swiglpk
swiglpk is a SWIG-generated glue layer to call the GLPK library.

Using it instead of the hand-coded glpk.cpp lets us drop the Boost,
Boost.Python, and Boost.Numpy dependencies, and indirect
dependencies on particular library versions (an issue for migration to
Sherlock 2?), the requirement to use a Python installation that has a
shared library libpython.so, and also delete the only .cpp file in
wcEcoli/.

**TODO:** Debug why the tests fail.

**TODO:** Pass GLPK pointers to ndarray data rather than copying to
an intArray or doubleArray.

**TODO:** Performance test and comparison with the old code.