Adjust the Firetasks to run the Plot classes
And move the 4 intermediate level base classes up to their parent directory so
the Firetasks won't list them when listing analysis plots to run.

That code should get replaced with explicit lists of subclasses, so the plots run
in a consistent order, to make it easy to run specific ones first, and to make it
easy to skip some of them.