control analysis plots sequence; reuse manual analysis CLIs for indiv plots
* Each `models/ecoli/analysis/*/__init__.py` defines a list of ACTIVE plots to
run, in order. This enables repeatable results to help debug the memory
problems. Edit this list to run some plots first or to comment out some
plots, whether in development or to check in.
* _Remember to add each new analysis plot to the ACTIVE list._
* The analysis Firetasks run the plots in that order rather than listing the
directory -- unless passed a list of plot names to run.
* Simplify the command line interface to run individual analysis and fill out the
argument defaulting feature by reusing the manual runscripts.
* `runscripts/manual/analysisCohort.py` et al now accept an optional list of
analysis plots to run.