Fix #377 script to diff 2 sim outputs
The aim is just to check if two simOut directories have identical content, e.g. after an optimization that's not supposed to change results. If the outputs don't match, this identifies which Table Columns and Attributes differ.

This ignores the Daughter1 and Daughter2 directories.

Q1. Should it compare the Daughter inherited state directories?

Sample output:
```
$ python runscripts/debug/diff_simouts.py out/manual/wildtype_000000/000000/generation_000000/000000/simOut/ out/experiment/wildtype_000000/000000/generation_000000/000000/simOut/
Using simOut out/manual/wildtype_000000/000000/generation_000000/000000/simOut/ to check out/experiment/wildtype_000000/000000/generation_000000/000000/simOut/

{'Table: EvaluationTime': {'Unequal Columns': ['calculateRequest_times',
                                               'calculateRequest_total',
                                               'evolveState_times',
                                               'evolveState_total',
                                               'merge_times',
                                               'merge_total',
                                               'partition_times',
                                               'partition_total',
                                               'updateQueries_times',
                                               'updateQueries_total']},
 'Table: Main': {'Unequal Attributes': [u'endTime', u'startTime']},
 'Table: RnaDegradationListener': {'Unequal Columns': ['DiffRelativeFirstOrderDecay']}}
```

Q2. The `EvaluationTime` differences and the `endTime` and `startTime` differences are fair. Should the script ignore this table?

Q3. What's up with `RnaDegradationListener/DiffRelativeFirstOrderDecay`? Is that a bug in psuedo-random number handling?

Q4. The `UniqueMolecules` table has no columns, just a single attribute `"collectionNames"`. Is it obsolete?