bring the agent command line & instructions up to date so I understand them again
* Update the command line parser and make the help more readable, e.g. via `argparse.RawDescriptionHelpFormatter` for manual line breaks in the help text.
* Add arg defaults or make them required to avoid obscure results.
* Construct the argparse after the subclass adds agent_types so it can set them as `choices`.
* Update some README instructions. More needed.
* Remove unused imports.

This should simplify the 99% case with only one environment agent. For now, at least print the environment agent ID clearly so we can paste it back into the trigger command.

This needs end-to-end error handling, e.g. if you try to add an agent to an incorrect OUTER_ID, it should tell you and not create the agent.

Should we update or delete some of the agent/ code like shepherd.py?