separate BLOBs from the Kafka message dict
In order to distribute cell agents across multiple machines, the daughter cell inherited state data has to get transmitted from one server to another (unless they have some kind of shared file server).

The inherited state data is a 250K BLOB (binary large object) and unsuitable to string-escape in JSON in the Kafka message payload, or to print to a log, or to put in a command-line arg when exec'ing a new agent.

So this changes our Kafka message payload format from a JSON string to a list of chunks. A JSON chunk carries the JSON dictionary and separate BLOB chunks carry any BLOB values. This is readily extensible, so e.g. we could compress the JSON data into a new chunk type.

Coming up, the cell agents will send inherited state data as BLOBs rather than pathnames.

But first, the Lens and Shepherd Clojure code has to support this new message format.