shrink the docker image
Getting some experience with Docker:
* Omit the git repo and such, shrinking the resulting Docker image from 1.67GB to 1.52GB.
* Use python3-compatible `except` syntax.
* Document the optional-CPLEX story.
* .gitignore .idea/inspectionProfiles/profiles_settings.xml

TODO: Update to openblas 0.3.4+ to get the threading bug fixes or to 0.3.5+ for additional performance improvements.

TODO: Build a container image with just WCM's requirements (no WCM code) and upload it to a private Docker registry in Google Container Registry for sharing. It currently takes 35 minutes to build that layer even when the result matches a locally cached layer. Build a second image on top of it that adds WCM code, compiled Cython, and Parca output to deploy for runSim. CI or Google Cloud Build could automate this.

FYI: There are more ways to shrink the Docker image like squashing layers and copying files to a new image without the Cython and C compilers. Some other month.

Q. Why set the parca output directory to `wholecellcontainer` in the example command line?