install openblas v0.3.5 instead of 0.3.3
... since that fixes threading bugs which cause inconsistent results and it also brings performance improvements.

App a technique found on the web to keep openblas' intermediate files out of the Docker images. It wasn't a priority but I couldn't think of any reason not to do it while adding code to manually build openblas, nor a reason not to apply it to GLPK.

Setting multiple ENV variables in one step also saves layers. That should save build time.

Combined, the wholecell Docker image went down another 390MB to 1.13GB.