put in the dual-daughters loop but not the feature
Put in `for l in [0]:` for daughter cell number `l` so the relevant code is properly indented in a `for` loop. Adding multiple daughters will then be a small code change.

Before adding the dual-daughters feature I want to test things out well and hear that someone wants to use it.