let gcloud launch workers in parallel
This is simpler and seems to be quicker, and we could add the `--async` option to make it return right away.

Also add a label `role=sisyphus` to each worker so a Compute Engine monitoring chart should be able to filter for them. But I haven't yet deciphered the filtering syntax docs.