quick temporary workaround for #635
Seed 28 is fatal. It's just a random seed so use another.

The bug is sensitive to code and data changes, but for now all seeds in [0 .. 99] are now OK. Do run `cloud/build-wcm.sh` to make a docker image with this new code before running it in GCE (Google Compute Engine).

With `self._seed += 1000` sim 28 gets stuck in a loop after time 1757.47! It doesn't even respond to ^C! Stuck in a C library?