add a Gaia test for python exception handling
In issue #635 , when the simulation task raised an exception and exited, Sisyphus uploaded the stderr (including stack trace) then Gaia kept rerunning the failed task on other workers. This adds a small test case to see if it's just the exception that causes that behavior and to speed up debugging.