make /.theano and /wcEcoli/fixtures writeable
Directories `/.theano/` and `/wcEcoli/fixtures/` must be world-writeable to be able to run WCM as any user in the container.

Typically it'd be `/home/.theano` but there's no home dir in the container for the host uid user.

TODO: Try to configure Theano to improve this.

TODO: Test if the Theano site-package really needs to be world-writeable. Maybe it's just `/.theano`.

TODO: Make Parca resilient to not being able to create `/wcEcoli/fixtures/endo_km/`.