cast the result of interpolate.splev() to float
`interpolate.splev()` is declared and documented to return `ndarray or list of ndarrays` although it returns a float in this case!

So appease PyCharm's type checker. Casting to float is clear approach with no runtime impact except a no-op function call. Assigning via a variable of type `Any` is a tiny bit cheaper but no so clear. Or we could actually apply `float()` to the result like `growthRateDependentParameters.py` does.

Adding to the mystery, mypy doesn't warn about this at all. It's good to have more than one static analyzer.