handle the remaining DictReader/Writer cases
This adds a simpler way to handle csv I/O including DictReader and DictWriter, then uses it to handle the remaining cases.

(We could use it to simplify `spreadsheets.py`, but that'll get simpler once we remove the PY2 code.)

Freeze on `bokeh==1.4.0` since we can make PY2 and PY3 virtualenvs both use it and adapt the calling code.