wcm-runtime: default to NO_AVX2 instructions
OpenBLAS runs faster with AVX2 instructions, but the impact on a cell sim in Docker is 7% in a quick comparison and maybe below the measurement noise.

Contrarywise, enabling AVX2 instructions means OpenBLAS won't build or run properly on Docker-for-Mac, which is not worth the debugging tarpit nor the complexity of having to know when to turn it off when building a Docker Image.

The bigger down side of disabling AVX2 instructions is it means we have to install OpenBLAS, numpy, and scipy from source, and somehow this seems to install cvxpy from source, and all that takes maybe 50 minutes.

To be clear, this is only when building the wcm-runtime Docker Image.

The build scripts still enable AVX2 instructions when building locally on Linux, since it's unlikely we'll run that image on Mac.

Record the OpenBLAS version number in a Docker ENV variable so it's visible via `docker inspect`, and "echo" some computed build options into the build log. (The echo doesn't happen when that build layer gets retrieved from cache.)

I tried to record more build options into ENV or LABEL fields for the sake of `docker inspect` but it appears to be blocked by a Docker security design decision!