refine augment-pycharm-package-list.sh
There's a new PyCharm release so we need to patch its list of packages where the import name doesn't match the PyPI name.

This change should handle PyCharm on Linux as installed by `snap` and makes it easier to support more OSs and PyCharm installers -- just add more directories to look in.

It's now careful to be idempotent.