update the mypy virtualenv to 3.9.0
The previous version of mypy fails on some f-strings with a misleading error message like this:

    --- Running mypy for Python 3.8 in wcEcoli3 ---
    runscripts/cloud/util/multigen_analysis_example.py: error: invalid syntax  [syntax]
    Found 1 error in 1 file (checked 466 source files)

The fix is to upgrade to mypy==0.790 on Python 3.9.0. I did that for the `mypy` virtualenv on Sherlock. It also has a fix in mypy's handling of `IO[]` types.

If you encounter these problems, rebuild your `mypy` virtualenv per the instructions in `mypy.sh`.