building wcm-runtime docker timed out...
It took 2h 15m so increase the timeout to 3h. That's not a lot of headroom but it's already inconveniently slow.

Soon I'll retest installing numpy and scipy from binary wheels, which is much faster and they now include the current OpenBLAS 0.3.13. The main reason to compile OpenBLAS, numpy, and scipy from source was to build with AVX2 instructions turned off.