Docker: optionally install OpenBLAS, numpy, & scipy from source
When building a `wcm-runtime` Docker Image, make the default be to install numpy & scipy from wheels with their embedded copies of OpenBLAS.

It's **much faster** and it's the usual approach so we're less likely to hit installer bugs like #1040.

The reason for compiling OpenBLAS from source was originally to get a bug fix and later to avoid the AVX2 problem building it in Docker-for-Mac.

This does change Parca outputs but that's all part of #931.