diff_simouts: ignore shell.log files
In fact, ignore all `simOut/*` files. It only compares tables, each of which is a `simOut/*` subdir.

(This calls `os.scandir(d)` which is faster and more convenient than `os.listdir(d)` since it gets the directory entries, making it easy to check which ones are subdirectories and to get the combined paths.)