send the lpad host info to fireworkers
When launching Fireworkers on GCE, send the Launchpad host info to override the `mongo-prime` default, and support `uri_mode`. This lets us manage the transition from to the `mongo2` host and support using Atlas servers.

Our local GCE MongoDB server is surely faster but probably not noticeably so relative to the workflow steps. Using Atlas would obviate the work and cost of maintaining our MongoDB server on GCE for Fireworks but not for Vivarium's use.