comparePickles tool to compare (dirs of) pickle files
Bug: Parca is creating different output every time it runs. (A problem using random numbers?)

To help debug it, this tool compares any pair of pickle files or a pair of directories of pickle files -- such as the `--intermediates-directory` containing output from all the stages of `runParca`.

It shows that the output has 9841 lines of differences in the very first stage, `sim_data_initialize.cPickle`, and the same number of differences carries through to the other stages.