Parca operons option
Add an option to enable/disable polycistronic operons in the Parca. This implements the simple, common, on-or-off case.

A later PR will add the harder case to request `both`, putting `simData` in a superposition of monocistronic and polycistronic operons, that is, running Parca both ways, writing two kb directories, and run the rest of the pipeline on both including in combinations with variants. That'll enable new variant analysis scripts to compare the impact of operons. It's much more complicated since a lot of code that assumes the `kb/` dir name will have to select between two dir names, the variant dir name pattern has to include the operon choice without breaking too much code, and we want to run the two Parcas in parallel.

Should the two kb directories be "kb/" and "kb2/"? "kb/" and "kb-poly/"? "kb/" and "kb/poly/"? Something else? The choice impacts wcm.py and (in the subdir case) borealis.

* Add the CLI option to runParca, fw_queue, and wcm.
* Record the operons_on boolean in `KnowledgeBaseEcoli.operons_on`.
* Assume that the `model_inspection` and `rnaDegRates` scripts should run with operons off. If anyone needs to run these scripts with operons on, then add CLI args to them. Do they get used enough to do it now?
* `Process`: It's broken to `raise warnings.warn()`. Call `warn()` and let it optionally raise an exception.