operons off/on/both in makeVariants
I'm trying the least disruptive approach: variant directories still look like `wildtype_000000` with a unique variant index to select one. When running with `--operons=both`, the additional set of variants start with `wildtype_100000`.

The idea is to set --operons the same in all pipeline steps: "both" in the Parca writes both mono- and polycistronic simData files while "both" in downstream steps uses both; "off" or "on" in the Parca writes one (mono or poly) simData file while "off" or "on" in downstream steps uses that file.

NOTE: In downstream steps the choice is really "both" or not; "off/on" do not select between mono/poly even if the Parca wrote two files.