fw_queue ability to run OPERONS=both
The fw_queue option `OPERONS=both` now runs the WCM workflow once with operons off and again with it on into an adjacent directory.

Set `OPERONS=both` in the PR build.

Q. Revert that before merging into master?
Q. Use the name suffix `'_poly'` or `'_operons'` for the Firetasks and the output directory?