import pickle instead of six.moves.cPickle
For #1311 Upgrade to Python 3.11, start taking out the `six` package for Python 2 * 3 compatibility, starting with importing `pickle` instead of `six.moves.cPickle`.

This one is complicated by the name change and by preserving `.cPickle` filename extensions.

It'd be a bit confusing to hack this change by changing only the imports (`import pickle as cPickle`), esp. since some of the code is already up on `pickle`.

Per PyCharm inspection warnings, remove unused imports and rename the local variable `complex` so it doesn't hide a built-in class.