rm testcase_npy_intp
Moved `testcase_npy_intp.pyx` to a separate little repo.

The NumPy team plans to fix the bug in 1.26.2. That could mean adding the int() calls to multinomial() or a broader change.

They try to change `RandomState` as little as possible, so the API change we hit was not intended. `RandomState` is there for compatibility but the recommendation is to use Generator in new code (since NumPy 1.17.0).