default to --no_jit
Fix #1400

Change the default to not use the Numba JIT compiler in the sim processes since the JIT compiler now takes several minutes to run and is not a net performance win.

This changes DEFAULT_SIMULATION_KWARGS which should change the manual runscripts, FW_QUEUE, and wcm.py.

ALSO replace `np.product()` with `np.prod()` since product() is deprecated and removed in numpy 1.26.